Makefile: Remove init target (it exists as a command now).
Because the command is moved from the build process to the xssh script
we need to generate a correct script file with the "DEFAULT_CONFIG" variable set and also include the config file in the install process.

Therefore we do a bunch of scripting aswell as copying the config file to the shared folder.